# openapi.yaml
openapi: 3.2.0
info:
  title: Train Travel API
  description: |
    API for finding and booking train trips across Europe, documented with OpenAPI 3.2.0 specification.
  version: 1.3.0
  contact:
    name: Train Support
    url: https://example.com/support
    email: support@example.com
  license:
    name: Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International
    identifier: CC-BY-NC-SA-4.0
  x-feedbackLink:
    label: Submit Feedback
    url: https://github.com/bump-sh-examples/train-travel-api/issues/new
servers:
  - url: https://train-travel-mock.train-travel-mock.workers.dev
    name: Mock Server
    description: |
      A stateful mock server for testing and development purposes. Book a trip end to end:
      ids returned by one call work in the next. Any `Authorization` header value gives you
      a private data space (none shares a public one); data expires after 24 hours idle.
    x-internal: false

  - url: https://api.example.com
    name: Production
    description: The production server for the Train Travel API
    x-internal: false
...
security:
  - OAuth2:
      - read
...
paths:
...
  /trips:
    get:
      summary: Get available train trips
      description: >
        Returns a list of available train trips between the specified origin and
        destination stations, departing from the given date and time (or from now),
        and allows for filtering by bicycle and dog allowances.
      operationId: get-trips
      tags:
        - Trips
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: origin
          in: query
          description: >
            The origin station: its ID, or a station name to search for (for example
            `Berlin` or `Berlin Hbf`). A name resolves to the best-matching station.
          required: true
          schema:
            type: string
          example: Berlin
        - name: destination
          in: query
          description: >
            The destination station: its ID, or a station name to search for (for
            example `Paris`). A name resolves to the best-matching station.
          required: true
          schema:
            type: string
          example: Paris
        - name: date
          in: query
          description: >
            The earliest departure, in ISO 8601 format. Without a UTC offset, it is
            read in the origin station's timezone. A date alone means the start of
            that day. Defaults to now.
          required: false
          schema:
            type: string
            format: date-time
          example: '2024-02-01T09:00:00Z'
        - name: bicycles
          in: query
          description: Only return trips where bicycles are known to be allowed
          required: false
          schema:
            type: boolean
            default: false
        - name: dogs
          in: query
          description: Only return trips where dogs are known to be allowed
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: A list of available train trips
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            RateLimit:
              $ref: '#/components/headers/RateLimit'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Wrapper-Collection'
                  - properties:
                      data:
                        type: array
                        items:
                          allOf:
                            - $ref: '#/components/schemas/Trip'
                            - $ref: '#/components/schemas/Links-Origin'
                            - $ref: '#/components/schemas/Links-Destination'
                  - properties:
                      links:
                        allOf:
                          - $ref: '#/components/schemas/Links-Self'
                          - $ref: '#/components/schemas/Links-Pagination'
              examples:
                GetTripsResponse200:
                  $ref: '#/components/examples/GetTripsResponse200ApplicationJson'
            application/xml:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Wrapper-Collection'
                  - properties:
                      data:
                        type: array
                        xml:
                          name: trips
                          wrapped: true
                        items:
                          $ref: '#/components/schemas/Trip'
                  - properties:
                      links:
                        allOf:
                          - $ref: '#/components/schemas/Links-Self'
                          - $ref: '#/components/schemas/Links-Pagination'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    page:
      name: page
      in: query
      description: The page number to return
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    limit:
      name: limit
      in: query
      description: The number of items to return per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      example: 10
  securitySchemes:
    OAuth2:
      type: oauth2
      description: |
        OAuth 2.0 following RFC8725 best practices, with Authorization Code
        and Device Authorization grants.

        On the mock server, no token is required: any value you enter gives you
        a private sandbox, and without one you share a public sandbox.
      oauth2MetadataUrl: https://example.com/.well-known/oauth-authorization-server
      flows:
        authorizationCode:
          authorizationUrl: https://example.com/oauth/authorize
          tokenUrl: https://example.com/oauth/token
          scopes:
            read: Read access
            write: Write access
        deviceAuthorization:
          deviceAuthorizationUrl: https://example.com/oauth/device
          tokenUrl: https://example.com/oauth/token
          scopes:
            read: Read access
            write: Write access
  schemas:
...
    Links-Self:
      description: The link to the current resource.
      type: object
      properties:
        self:
          type: string
          format: uri
    Links-Destination:
      description: The link to the destination station resource.
      type: object
      properties:
        self:
          type: string
          format: uri
    Links-Origin:
      description: The link to the origin station resource.
      type: object
      properties:
        self:
          type: string
          format: uri
    Links-Pagination:
      description: Links to the next and previous pages of a paginated response.
      type: object
      properties:
        next:
          type: string
          format: uri
        prev:
          type: string
          format: uri
    Problem:
      description: A problem detail object as defined in RFC 7807.
      type: object
      xml:
        name: problem
        namespace: urn:ietf:rfc:7807
      properties:
        type:
          type: string
          description: A URI reference that identifies the problem type
          examples:
          - https://example.com/probs/out-of-credit
        title:
          type: string
          description: A short, human-readable summary of the problem type
          examples:
          - You do not have enough credit.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem
          examples:
          - Your current balance is 30, but that costs 50.
        instance:
          type: string
          description: A URI reference that identifies the specific occurrence of the problem
          examples:
          - /account/12345/msgs/abc
        status:
          type: integer
          description: The HTTP status code
          examples:
          - 400
    Trip:
      description: A train trip.
      type: object
      xml:
        name: trip
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the trip
          examples:
            - ea399ba1-6d95-433f-92d1-83f67b775594
        origin:
          type: string
          description: The starting station of the trip
          examples:
            - efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
            - b2e783e1-c824-4d63-b37a-d8d698862f1d
        destination:
          type: string
          description: The destination station of the trip
          examples:
            - b2e783e1-c824-4d63-b37a-d8d698862f1d
            - efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
        departure_time:
          type: string
          format: date-time
          description: The date and time when the trip departs
          examples:
            - '2024-02-01T10:00:00Z'
        arrival_time:
          type: string
          format: date-time
          description: The date and time when the trip arrives
          examples:
            - '2024-02-01T16:00:00Z'
        operator:
          type: string
          description: The name of the operator of the trip
          examples:
            - Deutsche Bahn
            - SNCF
        price:
          type: number
          description: The cost of the trip
          examples:
            - 50
        bicycles_allowed:
          type: boolean
          description: Indicates whether bicycles are allowed on the trip
        dogs_allowed:
          type: boolean
          description: Indicates whether dogs are allowed on the trip
...
    Wrapper-Collection:
      description: This is a generic request/response wrapper which contains both data and links which serve as hypermedia controls (HATEOAS).
      type: object
      properties:
        data:
          description: The wrapper for a collection is an array of objects.
          type: array
          items:
            type: object
        links:
          description: A set of hypermedia links which serve as controls for the client.
          type: object
          readOnly: true
      xml:
        name: data
...
  headers:
    Cache-Control:
      description: |
        The Cache-Control header communicates directives for caching mechanisms in both requests and responses.
        It is used to specify the caching directives in responses to prevent caches from storing sensitive information.
      schema:
        type: string
        description: A comma-separated list of directives as defined in [RFC 9111](https://www.rfc-editor.org/rfc/rfc9111.html).
        examples:
          - max-age=3600
          - max-age=604800, public
          - no-store
          - no-cache
          - private
    RateLimit:
      description: |
        The RateLimit header communicates quota policies. It contains a `limit` to
        convey the expiring limit, `remaining` to convey the remaining quota units,
        and `reset` to convey the time window reset time.
      schema:
        type: string
        examples:
          - limit=10, remaining=0, reset=10
    Retry-After:
      description: |
        The Retry-After header indicates how long the user agent should wait before making a follow-up request.
        The value is in seconds and can be an integer or a date in the future.
        If the value is an integer, it indicates the number of seconds to wait.
        If the value is a date, it indicates the time at which the user agent should make a follow-up request.
      schema:
        type: string
      examples:
        integer:
          value: '120'
          summary: Retry after 120 seconds
        date:
          value: 'Fri, 31 Dec 2021 23:59:59 GMT'
          summary: Retry after the specified date
  responses:
    BadRequest:
      description: Bad Request
      headers:
        RateLimit:
          $ref: '#/components/headers/RateLimit'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            BadRequest:
              summary: A bad request response object
              dataValue:
                type: https://example.com/errors/bad-request
                title: Bad Request
                status: 400
                detail: The request is invalid or missing required parameters.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/bad-request
            title: Bad Request
            status: 400
            detail: The request is invalid or missing required parameters.
...
    Forbidden:
      description: Forbidden
      headers:
        RateLimit:
          $ref: '#/components/headers/RateLimit'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            Forbidden:
              summary: A forbidden response object
              dataValue:
                type: https://example.com/errors/forbidden
                title: Forbidden
                status: 403
                detail: Access is forbidden with the provided credentials.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/forbidden
            title: Forbidden
            status: 403
            detail: Access is forbidden with the provided credentials.
    InternalServerError:
      description: Internal Server Error
      headers:
        RateLimit:
          $ref: '#/components/headers/RateLimit'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            InternalServerError:
              summary: An internal server error response object
              dataValue:
                type: https://example.com/errors/internal-server-error
                title: Internal Server Error
                status: 500
                detail: An unexpected error occurred.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/internal-server-error
            title: Internal Server Error
            status: 500
            detail: An unexpected error occurred.
...
    TooManyRequests:
      description: Too Many Requests
      headers:
        RateLimit:
          $ref: '#/components/headers/RateLimit'
        Retry-After:
          $ref: '#/components/headers/Retry-After'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            TooManyRequests:
              summary: A too many requests response object
              dataValue:
                type: https://example.com/errors/too-many-requests
                title: Too Many Requests
                status: 429
                detail: You have exceeded the rate limit.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/too-many-requests
            title: Too Many Requests
            status: 429
            detail: You have exceeded the rate limit.
    Unauthorized:
      description: Unauthorized
      headers:
        RateLimit:
          $ref: '#/components/headers/RateLimit'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            Unauthorized:
              summary: An unauthorized response object
              dataValue:
                type: https://example.com/errors/unauthorized
                title: Unauthorized
                status: 401
                detail: You do not have the necessary permissions.
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://example.com/errors/unauthorized
            title: Unauthorized
            status: 401
            detail: You do not have the necessary permissions.
  examples:
...
    GetTripsResponse200ApplicationJson:
      description: A list of trips
      dataValue:
        data:
          - id: ea399ba1-6d95-433f-92d1-83f67b775594
            origin: efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
            destination: b2e783e1-c824-4d63-b37a-d8d698862f1d
            departure_time: '2024-02-01T10:00:00Z'
            arrival_time: '2024-02-01T16:00:00Z'
            price: 50
            operator: Deutsche Bahn
            bicycles_allowed: true
            dogs_allowed: true
            links:
              self: https://api.example.com/trips/ea399ba1-6d95-433f-92d1-83f67b775594
              origin: https://api.example.com/stations/efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
              destination: https://api.example.com/stations/b2e783e1-c824-4d63-b37a-d8d698862f1d
          - id: 4d67459c-af07-40bb-bb12-178dbb88e09f
            origin: b2e783e1-c824-4d63-b37a-d8d698862f1d
            destination: efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
            departure_time: '2024-02-01T12:00:00Z'
            arrival_time: '2024-02-01T18:00:00Z'
            price: 50
            operator: SNCF
            bicycles_allowed: true
            dogs_allowed: true
            links:
              self: https://api.example.com/trips/4d67459c-af07-40bb-bb12-178dbb88e09f
              origin: https://api.example.com/stations/b2e783e1-c824-4d63-b37a-d8d698862f1d
              destination: https://api.example.com/stations/efdbb9d1-02c2-4bc3-afb7-6788d8782b1e
        links:
          self: https://api.example.com/trips?origin=efdbb9d1-02c2-4bc3-afb7-6788d8782b1e&destination=b2e783e1-c824-4d63-b37a-d8d698862f1d&date=2024-02-01
          next: https://api.example.com/trips?origin=efdbb9d1-02c2-4bc3-afb7-6788d8782b1e&destination=b2e783e1-c824-4d63-b37a-d8d698862f1d&date=2024-02-01&page=2
...